Modified middleware to automatically logout user when logged on a different device.
Modified user and product to able to remove paranoically and restore them by trying tu duplicate them on the db.

+++ src/services/ProductService.ts
@@ -4,8 +4,9 @@ import ProductModel from "../models/ProductModel";
 import { IProductNew } from './../types/models/IProduct';
 
 import ElementNotFoundError from "../errors/ElementNotFoundError";
-import { ValidationError } from "sequelize";
+import { UniqueConstraintError, ValidationError } from "sequelize";
 import InternalError from "../errors/InternalError";
+import RuntimeError from "../errors/RuntimeError";
 
 export default class ProductService {
     private readonly REPOSITORY = new ProductRepository();
@@ -16,6 +17,14 @@ export default class ProductService {
             return user;
 
         } catch (error: any) {
+            if (error instanceof UniqueConstraintError) {
+                try {
+                    const restoredProduct = await this.restore(error.fields.name as string);
+                    return restoredProduct;
+                } catch (error: any) {
+                    throw error;
+                }
+            }
             if (error instanceof ValidationError) throw error;
             throw new InternalError(error.message);
         }
@@ -52,6 +61,9 @@ export default class ProductService {
             const product = await this.REPOSITORY.getById(id, transaction);
             if (!product) throw new ElementNotFoundError(`Producto con el id ${id} no encontrado en la base de datos.`);
 
+            const hasCommitedStock = product.stockCommitted > 0;
+            if(hasCommitedStock) throw new RuntimeError('No se pudo eliminar al producto ya que cuenta con stock comprometido en alguna orden.');
+
             await product.destroy({ transaction });
             await transaction.commit();
 
@@ -60,6 +72,28 @@ export default class ProductService {
             await transaction.rollback();
 
             if (error instanceof ElementNotFoundError) throw error;
+            throw new InternalError(error.message);
+        }
+    };
+
+    public restore = async (name: string): Promise<ProductModel> => {
+        const transaction = await this.REPOSITORY.newTransaction();
+        try {
+            const product = await this.REPOSITORY.getByName(name, transaction, false);
+
+            if (!product) throw new ElementNotFoundError('El producto que estas intentando restaurar no existe.');
+            if (!product.deletedAt) throw new RuntimeError(`El producto con el nombre ${name} ya existe en la base de datos.`);
+
+            await product.restore({ transaction });
+            await product.save({ transaction });
+            await transaction.commit();
+
+            return product;
+
+        } catch (error: any) {
+            await transaction.rollback();
+            if (error instanceof ElementNotFoundError) throw error;
+            if (error instanceof RuntimeError) throw error;
             throw new InternalError(error.message);
         }
     };
